pyproject.toml
[tool.poetry]
name = "mock-server"
version = "0.1.0"
description = "Mock REST API server"
authors = ["Bitjockey <bitjockey@jackpoint.network>"]
license = "MIT"
packages = [
    { include = "mock_server", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.9"
Faker = "^14.2.0"
bottle = "^0.12.23"
click = "^8.1.3"
python-dotenv = "^0.21.0"
flask-restful = "^0.3.9"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
black = "^22.8.0"
ipython = "^8.5.0"

[tool.poetry.scripts]
mock-server = 'mock_server.cli:cli'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
